Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main(){
-    int n, bin=0,rem=0,place=1;
-    scanf("%d",&n);
+    long long int n, bin=0,rem=0,place=1;
+    scanf("%lld",&n);
     while(n){
         rem = n % 2;
         n = n / 2;
         bin = bin + (rem * place);
         place = place * 10;
     }
-    printf("%d\n",bin);
+    printf("%lld\n",bin);
     return 0;
 }
